tests/requirements.txt
flake8
coverage
django-discover-runner
